marcus-bikes-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
httpx==0.25.1
alembic==1.12.1
pyjwt==2.8.0
setuptools>=68.0.0 